==> notebooks/2.Feature_engineering.ipynb <==
# Lib

%load_ext autoreload
%autoreload 2

import pandas as pd

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 10000)


import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import sys
import os
import glob

if sys.platform=='win32':
    sys.path.insert(0,".\..\src")
        
elif sys.platform=='linux':
    sys.path.insert(0,"./../src")


from sklearn.model_selection import train_test_split
from utils import utils_ml, parameters


from ipywidgets import *

if sys.platform=='win32':
    !dir .\
        
elif sys.platform=='linux':
    !ls ./

# Reading

## para saber en que directório uno se encuentra
#
if sys.platform=='win32':
    !echo %cd%
        
elif sys.platform=='linux':
    !pwd

if sys.platform=='win32':
    df = pd.read_csv('.\..\data\dados.csv', sep=',', decimal=".")
        
elif sys.platform=='linux':
    df = pd.read_csv('./../data/dados.csv', sep=',', decimal=".")


## dropping column "o" becaise it has 78% missing. and this is a boolean columns
# df.drop("o", axis=1, inplace=True)

### Renaming and formatting columns

df.rename(columns={"a":"a_int",
                    "p":"p_boolean","n":'n_boolean',
                     "k":"k_float","e":"e_float","h":"h_float","l":"l_float","f":"f_float","m":"m_float","m":"m_float","b":"b_float","c":"c_float","d":"d_float","o":"o_obj",
                     'i': 'produto','j':'categoria_produto','g':'pais'}
       , inplace = True)

## Transform categorical features
df['n_boolean'] = df['n_boolean'].astype("category")
df['p_boolean'] = df['p_boolean'].astype("category")


## creating columns os date and hour
df['data'] = pd.to_datetime( df['fecha'].apply(lambda x: x[:10]) )
df['hora'] = df['fecha'].apply(lambda x: x[10:])


# Split train, test and validation datasets

seman = df.groupby("data").agg( 


    total_um = ("fraude",lambda x: sum(x) ),
    total_zero = ("fraude",lambda x: sum( (x+1)%2 ) ),
    total = ("fraude","count" )
    
     )


seman['%_cusum'] = (seman['total']/seman['total'].sum()).cumsum()

seman.reset_index()

>> Splitting dataset in validation and train-test samples

n = 5
## Selecting the train and test to model (aprox 86% of full dataset)
datas_train_test = np.sort( df['data'].unique() )[:-n]
df_train_test = df[df['data'].isin(datas_train_test)]
## selecting the last n dates to validate (aprox 14% of full dataset)
datas_validation = np.sort( df['data'].unique() )[::-1][:n]
df_validation = df[df['data'].isin(datas_validation)]

df_train_test.shape[0]/df.shape[0], df_validation.shape[0]/df.shape[0]

>> Splitting the sample to contruct the WoE transformation

X = df_train_test.drop(['fraude'], axis=1)
y = df_train_test['fraude']

## Spliting the sample
X_train, X_test, y_train, y_test = train_test_split( X, y, test_size=0.3, random_state=0, stratify=y )

## Creating the label column to save
X_train['fraude'] = y_train
X_test['fraude'] = y_test

## Saving 
X_train.to_parquet(f"{parameters.data_path}\\train_dataset.parquet", index=False)
X_test.to_parquet(f"{parameters.data_path}\\test_dataset.parquet", index=False)
df_validation.to_parquet(f"{parameters.data_path}\\validation_dataset.parquet", index=False)

## WoE transformation pipeline

cats_cols = utils_ml.list_subset_words( df_train_test.columns.tolist() , ["bool","cat",'obj','pais'])
numeric_cols = [x for x in df_train_test.columns.tolist() if df_train_test[x].dtype.name!="object" and df_train_test[x].dtype.name!="category" and x!="fraude" ]

features = cats_cols + numeric_cols
features.remove('data')

features

### Encoding

dff = df_train_test.copy()

for variable in features:

    label = "fraude"
    dtype_ = 'categorical' if dff[variable].dtype.name=='object' or dff[variable].dtype.name=='category' else 'numerical'
    monotonic = True
    save_path_pkl = '.\..\src\\features\\'

    ## creating the WoE transformation and saving in pickle object.
    dff = utils_ml.encode_woe_var(dff, variable, label, dtype_, monotonic = True, save_path_pkl = save_path_pkl )

features_bins_woe = utils_ml.list_subset_words( dff.columns.tolist(), ["bin"])
features_bins_woe

len(features_bins_woe)

# Saving

dff

## Saving 
dff.to_parquet(f"{parameters.data_path}\\train_test_dataset_tranformed.parquet", index=False)

# pd.read_parquet(f"{parameters.data_path}\\train_test_dataset_tranformed.parquet")

# WoE to train dataset

### Train dataset

@interact

def display_metrics(variable = features ):

    label = "fraude"
    dtype_ = 'categorical' if X_train[variable].dtype.name=='object' or X_train[variable].dtype.name=='category'  else 'numerical'

    display(utils_ml.WOE_IV(df = X_train , variable = variable, dtype_ = dtype_ , label = label, monotonic= True))

### Full dataset

@interact

def display_metrics(variable = features ):

    label = "fraude"
    dtype_ = 'categorical' if df[variable].dtype.name=='object' or df[variable].dtype.name=='category'  else 'numerical'

    display(utils_ml.WOE_IV(df = df , variable = variable, dtype_ = dtype_ , label = label, monotonic= True))

# Requirements

path_root = parameters.root
!pip list --format=freeze > $path_root/requirements.txt
