==> notebooks/3.Stability_var.ipynb <==
# Lib

%load_ext autoreload
%autoreload 2

import pandas as pd

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 10000)


import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import sys
import os
import glob
import joblib

if sys.platform=='win32':
    sys.path.insert(0,".\..\src")
        
elif sys.platform=='linux':
    sys.path.insert(0,"./../src")

from utils import utils_ml, parameters


if sys.platform=='win32':
    !dir .\
        
elif sys.platform=='linux':
    !ls ./

# Reading

## para saber en que directório uno se encuentra
#
if sys.platform=='win32':
    !echo %cd%
        
elif sys.platform=='linux':
    !pwd

df = pd.read_parquet(f"{parameters.data_path}\\train_test_dataset_tranformed.parquet")

df.shape[0]

seman = df.groupby("data").agg( 

    set_data = ("data",lambda x: list(np.unique(x)) ),
     total_dias = ("data",lambda x: len(list(np.unique(x))) ),
     total_um = ("fraude",lambda x: sum(x) ),
    total_zero = ("fraude",lambda x: sum( (x+1)%2 ) ),
    total = ("fraude","count" )
    
     )

seman['%_cusum'] = (seman['total']/seman['total'].sum()).cumsum()

seman.reset_index()

colunas_orig_and_week = utils_ml.exclude_words( df.columns.tolist(), ["cat","bin","WoE","produto","fecha","data","hora","week","fraude"] ) + ['categoria_produto']

len(colunas_orig_and_week), len(df)

colunas_orig_and_week

## groups

## Ordenando as datas disponíveis no dataset
datas = np.sort( df['data'].unique() )
## total de dias para cada grupo
days = 10
## Total de grupos
total_groups = int(len(datas)/days) if int(len(datas)/days) == len(datas)/days else int(len(datas)/days+1)
## Lista de arrays contendo os dias para cada grupo
ls_group = []

i = 0
for x in range(1,total_groups+1):

    # print(x)
    group = datas[i:i+days]
    ls_group.append(group)

    i = i+days

print(f"Total de grupos:", total_groups)

ls_group

# WoE by groups

label = "fraude"
dtype_ = 'categorical'
df_woe_ivs_ls = []

for n,variable in enumerate(colunas_orig_and_week, start=1):

    print(f"{n}/{len(colunas_orig_and_week)}",variable)

    ## carregando o modelo pretreinado de
    optb = joblib.load(f".\..\src\\features\\pipeline_binning__{variable}__.pkl")

    for n, group in enumerate(ls_group, start=1):

        df_filtred = df[df['data'].isin(group)][[variable,label]]
        df_filtred = utils_ml.encode_woe_var_transform(df_filtred, variable, optb )

        ## computting the WOE and IV to variable transformed with _cat in the end.
        df_iv = utils_ml.WOE_IV(df = df_filtred  , variable = f"{variable}_bins", dtype_ = dtype_ , label = label, monotonic= True, optimal_binning = False)
        df_iv['group'] = n


        df_woe_ivs_ls.append(df_iv)

df_woe_ivs = pd.concat(df_woe_ivs_ls).sort_values("IV_var", ascending=False)

df_woe_ivs['Bin'] = df_woe_ivs['Bin'].astype("string")
df_woe_ivs.query("Bin!='Missing' and Bin!='Special'").to_parquet(".\..\data\\iv_woe_by_week.parquet")

## Data to explore in powerBI

# 3. Desempenho


- Como você pode garantir que o desempenho do modelo no laboratório vai
ser um proxy para o desempenho do modelo em produção?



# Requirements

path_root = parameters.root
!pip list --format=freeze > $path_root/requirements.txt
