==> notebooks/5.Modelo_analisis_negocio.ipynb <==
# Lib

%load_ext autoreload
%autoreload 2

import pandas as pd

import warnings

## ignore warning log in pandas
from pandas.core.common import SettingWithCopyWarning
warnings.simplefilter(action="ignore", category=SettingWithCopyWarning)

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 10000)


import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import joblib
import sys

if sys.platform=='win32':
    sys.path.insert(0,".\..\src")
        
elif sys.platform=='linux':
    sys.path.insert(0,"./../src")

from utils import utils_ml, parameters, metrics_evaluation, Pipeline_functions


from ipywidgets import interact

if sys.platform=='win32':
    !dir .\
        
elif sys.platform=='linux':
    !ls ./

# Reading

## para saber en que directório uno se encuentra
#
if sys.platform=='win32':
    !echo %cd%
        
elif sys.platform=='linux':
    !pwd

!dir .\\..\data

df_origin = pd.read_parquet(f"{parameters.data_path}\\dataset_structed.parquet")

scores = [x for x in range(0,101)]

## 6. Cost Matrix and confusion Matrix

- Queremos encontrar o ponto de corte ideal para o modelo já treinado
apresentado na parte (Score, onde 0 (zero) é o risco mais baixo e 100 (cem) o
mais alto). Sabemos que MELI ganha 10% do valor de um pagamento
aprovado corretamente e a cada fraude aprovada perdemos 100% do valor do
pagamento.Defina o ponto de corte que maximize os lucros do MELI, de
forma que todos os pagamentos com menor Score sejam aprovados e os
maiores sejam rejeitados.

    -  Além da matriz de confusão, uma matriz associada as vendas dado as predições feitas pelo modelo
    que resultaram ou não em fraude.

        - Para a matriz de **confusão** foram selecionadas as métricas:

        \begin{align*}
        Recall =​ \frac{​TP​​}{TP+FN}​​, && && FPR = \frac{FP}{TP+FP} \\
        \end{align*}

        - Para a matriz de **custo** foi criada a métrica:

            Custo     | PRED 0         | PRED 1
            --------- | ------         | ------
            Real 0    | gain           | lost by bloq
            Real 1    | lost by fraud  | Fraud (No lost)
           

        \begin{align*}
        &Ganho\_sem\_bloq =​ 0.1*gain-lost\_by\_fraud \\
        &Ganho\_com\_bloq =​ 0.1*gain-(lost\_by\_fraud+0.1*lost\_by\_bloq) \\
        \\
        &gain:= \text{ Ganhos por acertos do modelo nas campras que não foram fraude. (TN ticket) } \\
        &lost\_by\_fraud:= \text{ Perdas por o modelo não acertar as compras que foram fraude. (FN ticket)} \\
        &lost\_by\_bloq:=\text{ Perdas por ter bloqueado a compra dado que era uma compra legítima. (FP ticket)}
        \end{align*}
        
        O ponto de corte que maximiza os ganhos do MELI é definido com aquele ponto onde as métricas de "Ganho" tenham um máximo absoluto.

        Para a métrica de $Ganho\_com\_bloq$, que inclue no cálculo da perda de oportunidade daqueles pedidos que eram legítimos mas que foram bloqueados, o ponto de corte encontra-se em 86.

        Para a métrica de $Ganho\_com\_bloq$, que não inclue no cálculo da perda de oportunidade daqueles pedidos que eram legítimos mas que foram bloqueados, o ponto de corte encontra-se em 76.



@interact

def slice_score_in_matrix_confusion(score = (0,100), option = [False, True]):


    confusion_matrix, confusion_matrix_monto = metrics_evaluation.cross_tab_matrices(df_origin, cutoff= score, total = option)

    ######################### Métricas ##################################################
    TP, FP, TN, FN, P, N = metrics_evaluation.confusion_matrix_2_TP_FP_TN_FN( confusion_matrix, total = option  )
    gain, lost_by_bloq, lost_by_fraud, success_by_fraud, total_sell = metrics_evaluation.confusion_matrix_2_bussines(confusion_matrix_monto, total = option)

    Recall = TP/P
    FPR = FP/(FP+TP)


    ganho_com_bloq = (0.1*gain-lost_by_fraud-0.1*lost_by_bloq)
    ganho_sem_bloq = (0.1*gain-lost_by_fraud)


    print("\n\n","========================== Confusion Matrix =======================================")
    display( confusion_matrix )


    print("Total fraudes: ", P)
    print("Total Não fraudes: ", N)
    print("\n\n")
    print("Recall: ", Recall)
    print("FPR: ", FPR)


    print("\n\n","========================== Ticket confusion Matrix =======================================")
    display( confusion_matrix_monto  )

    print("Total vendido: ", total_sell)
    print("\n\n")
    print("ganho_com_bloq: ", ganho_com_bloq)
    print("ganho_sem_bloq: ", ganho_sem_bloq)


# PLOT

dic_gain = {"score":[], "TP":[], "FP":[], "TN":[], "FN":[],"P":[], "N":[],
             "gain":[],"lost_by_bloq":[], "lost_by_fraud":[], "success_by_fraud":[],"total_sell":[],
             "Recall":[], "FPR":[],
             "ganho_com_bloq":[], "ganho_sem_bloq":[]}

for score in range(0,101):

    TP, FP, TN, FN, P, N,\
    gain, lost_by_bloq, lost_by_fraud, success_by_fraud, total_sell,\
    Recall, FPR, ganho_com_bloq, ganho_sem_bloq = metrics_evaluation.gain_value( df = df_origin, score = score )

    dic_gain["score"].append(score)

    dic_gain["TP"].append(TP)
    dic_gain["FP"].append(FP)
    dic_gain["TN"].append(TN)
    dic_gain["FN"].append(FN)
    dic_gain["P"].append(P)
    dic_gain["N"].append(N)

    dic_gain["gain"].append(gain)
    dic_gain["lost_by_bloq"].append(lost_by_bloq)
    dic_gain["lost_by_fraud"].append(lost_by_fraud)
    dic_gain["success_by_fraud"].append(success_by_fraud)
    dic_gain["total_sell"].append(total_sell)

    ## Métricas de modelo
    dic_gain["Recall"].append( Recall )
    dic_gain["FPR"].append(FPR)

    ## Métricas de negócio
    dic_gain["ganho_com_bloq"].append(ganho_com_bloq)
    dic_gain["ganho_sem_bloq"].append(ganho_sem_bloq)


!dir .\..\\data

df_gain = pd.DataFrame(dic_gain)

path_save = os.path.join(parameters.data_path,"cutoff.parquet")
df_gain.to_parquet(path_save, index=False)
# df_gain

import plotly.express as px

fig = px.line(df_gain, x="score", y=["ganho_com_bloq","ganho_sem_bloq"], title='ganho vs score')

fig.update_layout(
    autosize=False,
    width=600,
    height=600,)

fig.show()

# Requirements

path_root = parameters.root
!pip list --format=freeze > $path_root/requirements.txt
